[NTOS] - Multiple ULONG -> SIZE_T fixes - Use SLIST_ENTRY, not SINGLE_LIST_ENTRY in InterlockedPushEntrySList, they are not the same! - Fix InterlockedCompareExchangePointer usage - _1MB is 1024 KB, not 1000 - Rename PDE_SIZE to PD_SIZE to reflect it's purpose (size of a page directory, not a page directory entry) - Add PD_COUNT, PDE_COUNT and PTE_COUNT for amd64 (should rather use PTE_PER_PAGE etc and move them to arch specific files. - Some more PFN_TYPE -> PFN_NUMBER
svn path=/branches/ros-amd64-bringup/; revision=46404

+++ reactos/ntoskrnl/mm/ARM3/contmem.c
@@ -391,7 +391,7 @@ MmAllocateContiguousMemorySpecifyCache(IN SIZE_T NumberOfBytes,
  */
 PVOID
 NTAPI
-MmAllocateContiguousMemory(IN ULONG NumberOfBytes,
+MmAllocateContiguousMemory(IN SIZE_T NumberOfBytes,
                            IN PHYSICAL_ADDRESS HighestAcceptableAddress)
 {
     PFN_NUMBER HighestPfn;
